wip: reset flow when loading a new chat/report
